move the independent server to the cmd
