+++ cmd/dashboard/adapter/manager.go
