+++ cmd/dashboard/adapter/redirector.go
