+++ cmd/dashboard/distro/.gitignore
