+++ cmd/dashboard/distro/placeholder.go
