+++ cmd/dashboard/keyvisual/input/core.go
