+++ cmd/dashboard/uiserver/embedded_assets_rewriter.go
