+++ cmd/swagger/docs/.gitignore
@@ -1,0 +1,3 @@
+/*
+!/.gitignore
+!/placeholder.go
