--- docs/swagger/placeholder.go
@@ -1,1 +1,0 @@
-package swagger
