+++ tests/dashboard/service_test.go
@@ -25,7 +25,7 @@ import (
 	. "github.com/pingcap/check"
 	"go.uber.org/goleak"
 
-	"github.com/tikv/pd/pkg/dashboard"
+	"github.com/tikv/pd/cmd/dashboard"
 	"github.com/tikv/pd/pkg/testutil"
 	"github.com/tikv/pd/server"
 	"github.com/tikv/pd/server/config"
